fix(041): comprovante de pagamento aceita centavos
event_payments.amount era Integer e arredondava os centavos digitados
(10.996,25 virava 10.996), deixando um falso 'Falta R$ 0,25'. Coluna
migrada para Numeric(12,2) (mesma precisão de sale_value); parse com
parse_brl (Decimal) em adicionar/editar/criar evento.

Migration: l8a9b0c1d2e3 (valores antigos convertem sem perda).

+++ app/models.py
@@ -401,7 +401,7 @@ class EventPayment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     event_id = db.Column(db.Integer, db.ForeignKey("calendar_events.id"), nullable=False)
     file_path = db.Column(db.String(300), nullable=False)
-    amount = db.Column(db.Integer, nullable=True)
+    amount = db.Column(db.Numeric(12, 2), nullable=True)  # aceita centavos
     created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
 
     event = db.relationship("CalendarEvent", lazy=True)
